infrastructure/gateway/requirements.txt
flask
requests
pyjwt[crypto]
